pyproject.toml
[tool.poetry]
name = "tna"
version = "0.1.0"
description = ""
authors = ["Dan Bentley <dan@numiko.com>"]

[tool.poetry.dependencies]
python = "^3.8"
wagtail = "2.12"
requests = "^2.25.1"
pyquery = "^1.4.3"
requests_cache = "^0.5.2"
psycopg2 = "^2.8.6"
gunicorn = "^20.0.4"
platformshconfig = "^2.4.0"
elasticsearch = "^7.11.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
django-extensions = "^3.1.0"
ipdb = "^0.13.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
